Fix dnd dependency bug
There was a super tricky bug where the wrong item was being passed
from the drag to the drop component. Turns out that the data in the
drag component is memoized, and you need to specify a dependency
array to update the memo when those variables in the dependency array
change value.