Add react router
Installed react router. Created stand in add and edit pages.
Had to reorganize the file structure to accomodate the router.
Functionality of the launch page stayed the same.

Fixed a small bug when opening links in the current tab. The callback
gets an array of tabs, not a single tab.