Use global state for links with react hooks
The prop drilling was getting pretty complicated. After exploring a few
different hooks patterns, I found this one that uses useReducer in
combination with useContext to create a global state accessible and
updatable by any component that chooses to access it. It's type safe
too!

The app is in a working state at this point, and the business logic is
now well separated from the UI. There is plenty of cleanup to do still,
but this marks a significant reduction in complexity.

Relevant React docs
https://reactjs.org/docs/context.html
https://reactjs.org/docs/hooks-reference.html#usecontext
https://reactjs.org/docs/hooks-reference.html#usereducer

Inspirations
https://www.thisdot.co/blog/creating-a-global-state-with-react-hooks
https://codeburst.io/global-state-with-react-hooks-and-context-api-87019cc4f2cf

Giving types to the dispatch function
https://stackoverflow.com/a/59432211/12671411