Use Chakra for AddLinkPage
The color mode is now controlled by Chakra, and accessible anywhere via
hooks.

The addlinkpage no longer uses styled components. The UI is a little
different now, but much better for the most part.

We are still using the themes/themes.ts file for a couple light/dark
color determinations. It would be nice to move away from this pattern
eventually and rely on Chakra's built in capabilities to achieve the
same effect.