Convert icon buttons to Chakra
Converts the action bar buttons and the add link page back button to
chakra.

The header of the add link page is still a bit hacky, with setting the
position of the back button as absolute. I'm struggling to find a way to
keep the back button on the left side of the header while centering the
title text on the page itself, no matter how wide the button is.