Refactor cards with Chakra
The card and card images now use chakra. This made it much easier to
expand the title of a card when hovered, resolving #1.

This introduces a bug, though, where the hover state of the original
cell persists when a card is dragged out of the cell. (it's really the
hover state of the card, but it remains in the original cell). I think
this could be fixed by creating a global context about whether a drag
event is in progress. If that is the case, then no cards should have
their hover states activated (ie _hover and _active should have
transparent background color). This solution requires a lot of
infrastructure, but I can't think of a better one ATM.

Closes #1