Bugfix: Spread object before key->val literals
TIL (the hard way) that the object spread syntax will overwrite any
previously defined keys in the object that is being defined. In this
case, we set the 'links' key of the return value, then used object
spread of the prevState, which overwrote 'links' with the links from the
previous state. This of course was a bug, because the state wasn't being
properly updated, and thus the changes would only be visible the next
time the state was loaded from storage.