Simplify drag event logic, hide last card on drag
When a global drag event is in progress, we can actually detect that
directly through the DnD hooks instead of storing it in our custom
global state. This simplifies logic quite a bit.

We can also leverage that to hide the last card when a card is being
dragged over empty cells.

We hide the last card if:
* the dragged item is over the cell (like every other cell, nothing
  special here)
* OR there is a dragged item, the card in the cell is the last card in
  the list, and the dragged item is the same as the card in this cell

This is a nice UI feature that was previously difficult to implement.