Fix: Bug where last card could not be dragged
As of the last commit there was a bug where the last card would
immediately disappear when dragged and could not be moved to a new cell.

I did not figure out the cause of the bug, but the workaround is to have
the Grid manage when a hover over an empty cell is in progress, and the
Cells update that state accordingly.