refactor: Allow nonstandard link urls
Allows users to create urls that don't follow the typical http format.

This is considered a feature, and not a bug, because certain strings
like 'chrome://newtab/' are handled in a special way by the browser but
are rejected by the JS URL object.

This creates the potential for the user to enter a bad link url, but the
worst thing that can happen is that they are sent to a nonexistent page.
Hopefully it's obvious enough for the user to know what went wrong and
how to correct the mistake.