build: Auto reload dev build of firefox
Uses the web-ext package to hot reload changes to the extension when
developing with firefox.

This creates an untracked profile and launches a dummy browser
that dies when the script is killed.

Concurrently is used to simultaneously launch webpack to rebuild the
assets and to run web-ext to watch for asset changes and reload the
extension.