fix: Store each link individually in storage.sync
This fixes a bug where the links ran out of storage (max 8 kb) in
Firefox due to large image favicon urls (1-2 kb each).

Now, links are each stored as their own item in storage.sync. A list of
identifiers (keys) is used to find all the links in storage, and the
order of this list determines the order of the links in the UI.

A function has been added to port links stored in the legacy format over
to the new storage format. The legacy links are not deleted, just in
case there's a bug and we need to recover them again.